+++ pages/DoctorEarlyAccess.jsx
@@ -1,6 +1,6 @@
 import { useRouter } from "next/router";
 import { NavBar } from "../components/home/NavBar/NavBar.jsx";
-import ProfessionalsForm from "../components/landingpage_extras/ProfessionalsForm.jsx";
+import ProfessionalsForm from "../components/landingpage_extras/DoctorEarlyAccess.jsx";
 import styles from "../styles/Home.module.css";
 
 export default function comingSoon() {
